Uniformize card html layout
Use the same html layout between agiledashboard and cardwall plugin.

This is a first step towards a unified look 'n feel of cards across
plateform.

TODO: externalize css of cards to the core of tuleap.
TODO: use mustache rendering?
TODO: do not compute the url of an artifact by hand