Tracker v5 Fix sla #1063: Dependency displays wrong field value
In tracker V3 (maybe in TV5 too) when a select box field values has a dependency to another one's, the displayed value is not always the real value in the DB, it's by default the first one in the select options.

This is due to JS since the value is displayed correctly when JS is totally disabled or the JS part in charge of field dependencies is commented.