Allow admin to replay a system event
If elasticsearch is down while tuleap tries to index a document, the
corresponding system event is flagged as error. The only way to reindex
the document (once elasticsearch is up) is to delete and recreate the
document in the docman.

Now the admin just has to click on 'replay this event' to reset the status
to NEW.