Another way to find the source artifact in the request
Needed when we create a story in a sprint.
We can choose an existing theme associated to the parent release of the
sprint.
But if we choose to create a new one, we should be able to choose a
theme associated to the product/corp/whatever parent of the release.