Update various shell scripts to be POSIX compliant
Debian's /bin/sh is dash and not bash and doesn't support all the features
that bash does support but it is POSIX compliant. It's thus important to
use only POSIX features when the shebang points to /bin/sh and not
/bin/bash.

In particular "function foo { ... }" is a bashism, the correct way to
create functions is "foo() { ... }".