Check on client side the size of the upload
This fix sla #1854: if the user exceed the max file size upload, then we
prevent them to submit the release, so that they do not loose their time
waiting for backend failing.

This can only be done client side as PHP cannot prevent the client to send
too big files.

Please note that it works only on browsers which support the html5 File
api[1]. An alternative would be to use a polyfill[2]

[1] http://caniuse.com/fileapi
[2] https://github.com/Modernizr/Modernizr/wiki/HTML5-Cross-Browser-Polyfills#file-api