Fix app user id in lxc
When you use Tuleap in a LXC, there may be chances that the user id of
the app user (codendiadm|www-data) diverges from the real owner of the
working files (you, the developer).

This script automates the step given in the following page:
https://tuleap.net/wiki/index.php?pagename=Development%20environment%2FVirtualBox&group_id=101#Sharing_files_with_host_with_lxc

    usermod -u youruid codendiadm
    groupmod -g yourgid codendiadm
    usermod -g yourgid codendiadm
    find / -uid yourolduid -exec chown codendiadm {} \;
    find / -gid youroldgid -exec chgrp codendiadm {} \;

It is a first step to automatize the launch of a ready-to-use
developer-friendly container.

Change-Id: Ib2587778381c02dcbce327c87ad8b825355ca9b8