Refactor ListValue json formatting
Previously JSon format for ListValues was used for 2 purposes:
- OpenList management
- Cardwall rendering & edition

OpenList came first and the JSon format was done to match the JS library
used for the rendering of the field (ProtoMultiSelect). Hence the format
included "value" and "caption".

Cardwall needed a JSon representation of ListValues too and tried to leverage
on JSon done for openlist, so "caption" started to percolate outside the
initial scope.

The purpose of this refactoring is to clearly separate concerns:
- One specific format for OpenList to match what is needed by ProtoMultiSelect
- One standard format used by Tuleap itself (Cardwall and TriggerRule builder)

The later is meant to be THE reference Json representation of a form element
and a list value.

Change-Id: Ia4076251661cf45adee483d1db0ebe231e90ee6e