Allow DB errors to be converted to Exceptions
Previously, DB errors emited notices and it was up to the developer to
care if the query was successfull of not.
Basically, if a query lead to an error (data correctness, server issue,
etc), if developer did a bad job of not taking this into account, the
execution flow continued (potentially with inconsistent data).

With this modification, once enabled, as soon as an error occurs, an exception
is thrown and if the developer didn't care to catch it, it aborts the current
work.

Change-Id: If20f1f02e5e7fda2c357144cd9cefa9827ac9d06