Send an array of triggers instead of a map
The collection of triggers is an associative array in php. When we send
the content in JSON to the client it is converted to a map ({…}).
However when the collection is empty, it is converted to an empty array
([]).

Therefore it is difficult to the emptyness of the collection in
JavaScript since we need to detect if we received an [] or a {…}.

Therefore always send the collection as an array.

Change-Id: I7e9a1e31688ff40a491b7dda896d62c2d0461fee