Inject dependencies
We need only once instance of the logger, therefore inject dependencies
instead of relying on various object (like Tracker) to build stuff.

There is still work to be done to clarify dependencies. For example is
it the role of the WorkflowFactory to return a RulesManager for a given
tracker. To be done in a next step to not clutter this changeset.

Change-Id: I3ec4e6d4be40b5a64bae3126cb48d6e707efe6b2