Search only on planning items
Given I have the following hierarchy:
  Product -> Release -> Sprint
and the following plannings:
 - Release planning
 - Sprint planning

Then I should only see release and sprint milestones.

Please note that the following configuration is NOT supported:
 - Product planning
 - Sprint planning
(no release planning)
The current behavior is to propose only sprint milestones, nevertheless
a warning will (in a near future) be displayed if we encounter such configuration.

Change-Id: I45e0d534d68d27742a57e84be5242b48ccd6dda2