Allowing a user to add themselves in gerrit groups if gerrit "set_account" command failed.
This is in response to the following issue:
If a gerrit server (2.5.2) is set-up to store data in a mysql database and to use ldap users then
it is impossible to automatically create accounts for ldap users on it. The user must first log-in
to the Gerrit UI for their ldap account to be recognised.

Unfortunately, if a user logs-in after their Tuleap user-group was copied over to Gerrit then they are
not listed as part of this group. The workaround is to allow a user to manually update themselves via
a button in their "account management" area.

Change-Id: Ia6f6d7e42eb4e5d7046ace0e0c207fefafe0d14b