Remove chef and vagrant stuff from repository
Vagrant and chef stuff clutter the source repository, there is no reason
they are so strongly coupled with it.
Therefore, this commit extracts those two elements in 2 dedicated repo:
* Chef cookbook: https://tuleap.net/plugins/git/index.php/101/view/157/
* Vagrant setup: https://tuleap.net/plugins/git/index.php/101/view/69/

There are 2 repositories because vagrant is one stuff for development
but chef cookbook can be reused outside the context of vagrant, to deploy
lxc machines or even productions servers.

Change-Id: I5438f1f464986830ea4048b262d1ea84a8547dec