request #5951 - Avoid hardcoded redirect on login when possible
When we configure Tuleap behind a reverse proxy that does URL & domain rewriting
most of the time, it doesn't play well with hardcoded redirect (with the domain
name in the redirect request).

As we only need the domain when we change of protocol (http <-> https), starting
this commit, Tuleap will do relative redirects (without domain name) when there
is no protocol change.

As a matter of fact, when Tuleap is configured in such a setup, it should stick
to one protocol (either full HTTP or full HTTPS) and not allow both.

Change-Id: I789d914b466d573453a6c33c8f7b4783f0c66bcd