Do not double escape urls in html text
There is a side effect. If the text is already containing a link to a
reference:

    <a href="…">art #123</a>

then the link to the reference will be injected inside the existing <a>:

    <a href="…"><a href="…">art #123</a></a>

The browser will interpret it as the following:

   <a href="…"></a><a href="…">art #123</a>

The first empty link will be hidden from the user, therefore we assume
that it is a good compromise instead of having either no links or
crappiness: fix request #5986 - URLs are corrupted by a double escape.

Change-Id: I2f46bcf14c5ea941ff807673ad41f416f48ac686