Generate changelogs of Themes
As of today, only plugins changelog is semi-auto generated. This is a
problem when theme is upgraded and not its version number. So now the
themes have their section in the changelog.

Special dedicac^Wnote for Experimental theme: it *includes* things from
common pseudo theme. If something is changed in common, the version
number of Experimental should be changed if we want the new stylsheets
to be built during package generation.

Change-Id: Ie62fa9291c584e0579d14a2f944caeac5ef6188a