Resurect FRS ghost files.
In between request #5808 and request #6338, changing a file from one package
to another or changing it's release date, make the file disappear.

Introduce a tool to make those files alive again.

It's not a forgeupgrade bucket because it needs to be executed by hand and
project per project to identify potential errors

Change-Id: I73c3441d9ac225d03ee39781ba773425fe004514