The gerrit driver is now given by a factory
Instead of injecting the gerrit driver everywhere, inject the driver factory
instead so that the driver used depends on the server (one server is
reachable through gsql while another will be REST compliant).

Future step: remove the dependency on the factory and ask to the gerrit
server what is its driver.

Change-Id: Ib855ef03d035f91933b34e6db9c4673ef71cbaa3