request #6521 - Burndown computation seems broken
Please note that we are not ranking with the date as this is not discriment
enough. If 2 changeset have the same timestamp (for instance with workflow
trigger) we end up with 2 matching changeset instead of one.
To avoid this situation, we order with changeset.id (unique).

Please note however that, ranking by changeset id, can be misleading if we
start to introduce changes in the past (the new changeset will have a newer id
but it's date might be before). 99% of the time it should be transparent.

Change-Id: Ibeadd6339e45ac5f51355cdffa04f510ef3ed193