Follow-up to tasks #6652: Change style of add/remove criteria
The issue is that when you edit a card in the agiledashboard, a
javascript error is raised and prevent the modal to submit the change.
Instead the modal disappear and the page says: agiledashboard does not
exist.

This come from a missing rename following the move of the dateTimePicker
in the core introduced in commit #481523c3b6c48e66eb4d78ad7af8c1d49051d447

Change-Id: I59b2cfb3e0978c7aaa66ca21dfa63973166cbebd