Introduce a "sample project feeder"
Some tools to activate plugins and create a sample project call guinea pig with
agile template a couple of fake data.

This will create 2 users (richard_cover and alice_tyrell) with welcome0 as
password.

This is meant to be used on a bare tuleap deployment at the time being
(mainly because it creates new users) but could be modified later on to be
imported in any existing instance.

Change-Id: I3b87e80dc6a67d0b89107889955db8961cdd302f