Display edit button in gitphp UI
This is a first step of story #6971: Git online commit in the Tuleap UI

As we have no control over gitphp code, the button is injected through javascript,
only if 1) the site admin activate the corresponding variable 2) the repository is
configured to allow online edition 3) the user has write access on the repository.

Please note that the 2) is hard coded, you will have to edit manually the code to
test the feature: set $result[self::ENABLE_ONLINE_EDIT] variable to true in GitDao line 492.
The ability for a repository admin to change this setting will be done in another
changeset.

The [edit] button does nothing and is disabled if the user is targeting a file in the past.

Change-Id: I1ca81e33d4834f10ebb4b2d68ecccbf15a6fab40