Add possibility to remove blockquotes in html answer
This is a follow-up of story #6301: reply to a follow-up comment by email
Part of tasks #6991: handle html format

We are able to remove all blockquote elements.

Furthemore, we can detect the citation header line (On XXX, Toto wrote:)
for mail coming from gmail. Other MUA that we tested (Apple Mail and
Thunderbird) do not identify clearly this sentence.

**Note:** This is sparta^Wnot used yet.

Change-Id: Id5fb513235e20f390754c1f1684c0f7dee49e001