Do not display empty pane box for project
When the project into you are searching for has no wiki, forumml, or tv3
service activated, the panel on the left is shown empty wich is quite
ugly.

No we display a (small) message to inform the end user.

Fix tasks #7101: The "Recherche dans <MyProject>" box may be empty.

Part of story #5756: search with nice interface

Change-Id: I80b0757533c7a7f0c1de551238a3ef9cf3939915