Add Tuleap welcome tour
In order to improve the onboarding user experience, the first time the
user connects to tuleap, we display a welcome message and provide a
basic tour to discover the main features (access to projects, edit
preferences, …).

Unless the user hit escape or click on the [End] button on the tour, the
popover will be displayed again the next time the user hits /my/.

We register some basic stats about usage for future improvements: which
step is shown and at which step the tour is ended.

Change-Id: I8145eab516116a599d3addfe9cfe0ed7a92ac7da