Remove legacy function group_getid_by_name()
Fix request #7211: Fatal error while accessing to a specific svn
revision.

This is a follow-up to commit #f1257e04bf4ccba9de6bba85aabc6038057f98d9
where we removed a bunch of require_once in order to save trees.
Unfortunately, there is edge case when legacy methods are used with an
obscur require chain. (We are far from the phpwiki spaghetti by the way)

The solution for this issue is to remove the legacy function and use the
dedicated object instead.

Change-Id: Ia4b45e55c85765d4689bb37b924271126ba84112