Fix request #7373: Remove "target" settings for link in CKeditor configuration
With the default CKeditor, we can set the target attribute for links
(eg: <a href="…" target="_blank">…</a>). This is an issue because this
attribute is forbidden for HTMLPurifier and there is an inconsistency.

Now when user click on the [link] icon, she cannot anymore set the
target attribute. The same for [image > link].

Change-Id: If03519b4352bbffa681ceda52a10efc4395bdbb8