request #7491: Table columns are not saved during XML tracker import
Beware, here be dragons.

During XML import, columns are created in session. However it lacks some
information like the field id, which is pretty normal since we do not
have an id during the import. This means that we cannot use getColumns()
method because it relies on the presence of the field id for each column
to be build.

Hacking getColumns() to return the internal collection of columns (with
the reference to the field object) is not an option since it breaks
things.

The trick here is to delegate to the temporary renderer object to inject
the columns in the real one.

(╯°□°）╯︵ ┻━suoᴉssǝs━┻

Change-Id: I9e2b9713ee91093cc42ee0c4919bc405a69407b8