Allow to define a directory for combined script outside sources
When one develop Tuleap with shared sources, there is a conflict of file
ownership (at unix level) between the host (where the files are stored)
and the guest (where it's executed).

Most of the time it's harmless as sources are read-only stuff.

But for javascript combined, it's not the case, the application might (and will
generate a file on the behalf of the user). If the application is not granted
write permissions on the destination directory, it will fail and user will not
have javascript (and save 1.6MB of bandwidth).

With this commit, developers can define an alternative directory for combined
and map it at apache level.

Change-Id: I5c13877295bb4182c351c0a96ed7fbb78b68c8a3