Refactoring: plugins should not leak into the core
If you pollute core with plugins stuff, then you have
to update both the core and the plugin(s). We really should consider the
core with extension point so that it is easier to add behavior for other
plugins.

OCP rulez!

The Full Text Search system events are now run as 'app' instead of 'root'

Change-Id: Ic5d76ff4bb795a05aa522a9f1b1a5db7b6fd26a6