request #7536 - Make git urls more customizable
Sometimes you don't have SSH access to your git repo, hence the UI should not
display this possibility.

Sometimes you need to setup SSH access on another port and end users should be
informed of that.

Most of the time docker and freaking reverse proxy are involved but that doesn't
really matters.

Change-Id: I4f476ba455ce4b0c2e7e722d36681d4d2c0f0b9e