Do not send layout_id in _GET and _POST
When the user add/update/remove a column in its custom layout, the form
has action="…&layout_id=21&…" and a post variable is sent with
layout_id=-1 (to indicate that we want to generate a new custom layout).

When PHP receives the request it will populate the _REQUEST variables
accordingly to php.ini variable (request_order and variables_order
(the later is specific to Tuleap, see pre.php)). In default
configuration, the order is GP: GET, then POST. This means that _REQUEST
will be filled with _GET variables, then with _POST variables. In our
example, _REQUEST['layout_id'] will be equal to -1. However, if one
update the php.ini / php.conf to change the order of variables to PG:
POST, then GET, _REQUEST['layout_id'] will be equal to 21.

The code that manage layouts delete old layout. This means that if
_REQUEST['layout_id'] is -1, it will create a new layout based, will
migrate widget from the old layout to the new one, and then will delete
the old layout. Here is the trick: if old layout = new layout then the
project references a layout that is deleted.

To avoid this situation, we do not send anymore the layout_id in the GET
so that there is no ambiguity and it decrease the chance of having a
blank dashboard.

Change-Id: I87ec5944ac84632b0bb801d76d218a217d2249da