request #7620 - Cannot add non LDAP user to project
Yes, it is not really beautiful, even a bit crappy.

Problem appears when you have a Tuleap user (non-ldap one) who has an
 homonym in your LDAP directory. When you add this user to a project,
 the LDAP plugin will find him in your LDAP and try to create an 
account for him but he already has one so the exception 
LDAP_UserNotCreatedException will be raised and the process of 
adding him to the project aborted.

A proper workaround would be to check upstream whether it is really a
 ldap user or not (PFUser class has such information) but the event 
used by the LDAP plugin does not seem to pass the user object thus 
preventing us to do this check beforehand.

This issue probably existed since a long time but without any impact 
as the check on the returned user prevented to launch any 
synchronization if LDAPPlugin wasn't able to create an account for
the user. Today, with the Exception raised, it still prevent 
synchronization but the throw also prevents the continued execution 
of the "add user to project" action.

Change-Id: I526e50be2157ea484e54e881bd55bb88af07b36d