request #7649: Admin can activate an account when user comes from ldap
By default, users coming from ldap are activated (status = A). This
behaviour is not acceptable in every environments.

Admin may want to validate every users. An option already exists to
define the default status (sys_ldap_default_user_status) but if you
choose 'P', and sys_user_approval is set to 1 (force admin approval)
then no email is sent to the administrator.

The aim of this contribution is to send an email to admin like it is
done on the regular account register form.

The modifications forced us to rewrite TokenAuthenticationTest to a rest
test in order to manage restler dependency.

Change-Id: I2468151f56bf2a75e710a8a3440cadb2194a8965