Request #7656 - Migration DocmanV1 fails when UGROUP_NONE or UGROUP_ANONYMOUS
Problem was coming from the fact that ugroup_db_get_dynamic_members can
in some cases (UGROUP_NONE and UGROUP_ANONYMOUS) return an empty SQL
query. It was not previously catched by the code leading to a fatal
error (particularly on import from DocmanV1). Empty sql query is now
catched and an empty data access result is returned in this case.

Change-Id: I60cb3b69bbdd842348ed46bff9df75a8dbe02c33