story #6855 - artifact copy: Recursive copy
XML export of an artifact has now a new field exported: artifact link.
This field expose only children of the artifact, other relations are
ignored.

The logic mostly revolves around a ChildrenCollector object. This class
is responsible of collecting in a stack the various children we
encounter during the xml export, starting from the root artifact (the
one that is copied).  Then, until this stack is empty, we continue to
export in xml the corresponding artifacts, à la breadth-first search.

This mechanism is then reused during the xml import: each child that is
present in the xml is added to a ChildrenCollector and with the help of
a mapping between original and new artifacts, we can add the children to
artifacts in a new changeset.

Notes:
If a child is not readable by the current user, then it will not be
exported in xml (thus not copied) as well as its own children.

Please note that the xml is flat: no hierarchy is shown, only artifact
links field links artifacts to each other. We should add semantic to the
link when we will extend the xml import to handle other types of
artifacts relationship.

In order to be able to import a hierarchy of artifacts, the xml must
contain reference to the tracker ids. It is of course optional when you
do not have hierarchy like in tv3 to tv5 migration.

Change-Id: I9a7f78bc7b57c57f71efa1a64cae9cc1c013357f