request #7816: Do not stop batch indexation if one particular file cannot be indexed
If one of my document is too big for elasticsearch to be indexed, an
exception is raised and it prevents indexation of other valid documents
when we index the whole project. Now it skips the error and raise only a
warning if some documents are not indexed.

Bonus: use WrapperLogger to add the [Docman] prefix instead of repeating
ourselves.

Change-Id: Ic56a48f2fc7320afa2591035464b3a74c9403439