request #7836: Fatal error while deleting a Wiki item
When you try to delete a Wiki item in the docman, it raises a fatal
error, because $user parameter is missing in the event reused to index
the wikipage.

This issue seems to appear only when a wiki page is referenced by two
different locations in the docman, I think.

Change-Id: I3908257bbe959fd3cc5ebfc3c461fab0267cf03c