No admin mass change view if no available mirror
If a project cannot access to any mirror, then admins should not be able
to masschange the mirroring for this projects (as of today they
encounter an empty list of mirrors and it is not the best user
experience ever).

If there is no mirror available for a project, this change prevents
access to the corresponding tab in `project » git » admin`.

Implements: story #7709 - activate mirroring per project
Change-Id: Ic792d92dae9f6f8c942a33d6c8e601afb95127e9