Editable headline
As a siteadmin, in order to quickly customize the homepage without
having to understand the arcanes of site-content and without having to
understand in which specific file I have to do my modifications, I want
to be able to edit the headline of the Tuleap homepage directly from the
web ui.

site admin » configuration » configure homepage

The admin can choose if the new (standard) homepage should be used. Note
that it will be displayed only on FlamingParrot theme.

Implements: story #7890: new tuleap home page
Change-Id: I9ad3f466c17e460531eefe2564cf9caf5160b195