Add BasicAuth method if no token/cookie provided
This task is part of story #7701 Add REST basic auth

You can test it through the Web UI or with a curl.

Web UI
------

First, without be logged in, go to projects/ -> GET /v1/projects -> it must asks for credentials.
Then, as a user logged in, go to projects/ -> GET /v1/projects -> it must lists projects

curl
----

First, test with

$> curl -XGET ".../api/v1/projects" -H 'Content-Type: application/json' -H 'Accept: application/json' -u username:password

This must lists projects for user.

Then, generate a token. Once done, list project with

$> curl -XGET ".../api/v1/projects" -H 'Content-Type: application/json' -H 'Accept: application/json' -H 'X-Auth-Token: token' -H 'X-Auth-UserId: user-id'

Change-Id: I70076994fd012a63bd7188f62ee156b681a0ab22