Editable homepage logo
Like the headline, site admin is now able to change the homepage logo of
her Tuleap instance by uploading it directly from the web ui.

This requires to deploy a new apache rules to be able to deliver images
stored in in a subdirectory of /var/lib/tuleap/ subfolder.

Implements: story #7890: new tuleap home page
Change-Id: I86793742b5e9437e38da2b11b593afd05166767e