Refactor Tracker XML export: replace adhoc Field and Values + filter export
Until now, field and values were generated with relative IDs
ie. IDs started at 1 at the beginning of XML export and were
incremented along the process.

With this commit, we now use internal ids (field and value
ids) that are already unique and integers.
This will allow us to make easy reference when doing cross
references without bothering to keep a mapping between DB
ids and export ids

In addition to that we now only export the "used" form elements as it was
source of conflicts with some fields imported two times.

Change-Id: I59899c1ca7bbf4757f48bc1873458638cf2638bb