Move sys_allow_anon in the database
So that it is easier to update the configuration and it prevents
inconsistencies (e.g. sys_allow_anon = 1 && sys_allow_restricted_users = 1)

If the variables are still present in the local.inc file, then a warning
is display to ask to site admin to update the file.

Change-Id: Ie9cca82102b0070bc9aba72f2e663f7905ce7885
Implements: story #7813: grant restricted users access to git repositories