story #7684: Add route phpwiki/:pageid/versions
All previous wiki routes and projects/:id/wiki were renamed to phpwiki/

This route retrive the raw content and the HTML formatted content of a PhpWiki page's revision.
I'm not proud of how this is done, but we didn't find a proper way to do this.

So ... I just want to say "sorry" in advance to the people that will read this code.

Change-Id: If2e11a1b0e2b15c454d59126dffbd8a933d2ad37