request #8151: Request-URI Too Large in GET /users_memberships
Passing long list of users in GET request is not possible when there is
a huge number of users to fetch. Since the feature is used as of today
only for gitolite, we use now a query field to select only users with
ssh keys. This solution should be future proof since we can later on add
other options like 'all_users', 'self', …

As retrieving self membership information since not be used, we drop it
in order to remove dead code.

Change-Id: I001b48584acf459f9dadbd4a096b8c2427d79b54