story #8168: Export and import list fields
For now, the open lists are not taken in account.

In this commit, we are able to export and import list fields like selectbox or multiselect box
from one platform to another. We support the three binds (users, ugroups and static).

In addition, we fixed a weird behaviour for the list fields bound to ugroups in the report table.

Change-Id: I72dbc9817e7973d22c3f4f7344537bfc98de29f0