request #8210: Default value not set when unsubmittable field is submitted
Given a selectbox with a default value
And I cannot submit,
When I POST an artifact with a value for this field
Then the created artifact should have the default value.

Another option would have been to reject the request (403). However I'm
not willing to break compatibility for clients that are relying on the
current behavior.

Change-Id: I93ff503f5e87c39fe69bfd302346c3b99a05de7e