request #8189: An user with an IPv6 must be able to log in
We do not need to use IP address to check is an user is authentified or not.
It is quite frequent now that thare is a lot of users using the same IPv4 with
things like Carrier-Grade NAT (aka CGN or LSN) or mobile networks.
An IP is not an interesting information to authenticate an user with, it is just
interesting to log it for audit purposes.

Change-Id: I272f5641c49e74fa8b638330cbe9435f212276f1