request #8254 - Have a way to override tuleap's permissions
You can add a PermissionsOverrider object in a
PermissionsOverrider.php file in /etc/tuleap/local_glue/
that implements the PermissionsOverrider_IOverridePermissions interface.

This class will be used to override Tuleap's decision regarding the
access of a user to the platform or the project. If such an object
has not been implemented, normal permissions will apply.

Change-Id: I009f22dbb35eee45e3ad3829ebc787ace4fd4eca