story #8170: export/import attachments
The project export and import process is now able to
export/import artifact attachments.

Please note that attachments are not historized.
We export attachments that are linked to the artifact in its last changeset.

This change has no impact on artifact copy and on tv3 to tv5 migration.

Be careful, we have to run this script as `codendiadm` in order to
have to good access rigths in the imported attachments.

Change-Id: I9252c97ebac5624089406e332ba679ada4aec897