story #8173 Use an archive in project export/import process
We are now using an archive to do the project export/import process.

We introduce a new option (-o) for export_project_xml.php to define
the path of the created archive.

A second option (-x) is added in order to print the project.xml content
in your terminal (debug purpose)

Example:

* src/utils/php-launcher.sh src/utils/export_project_xml.php -p 116 -u awesome_site_admin -t 1172 -f -o /tmp/my_archive.zip
* src/utils/php-launcher.sh src/utils/import_project_xml.php 258 awesome_project_admin /tmp/my_archive.zip

Change-Id: I207fa9ddbf77a5636d3d2125d2598694b6c15da8