Allow to define a path for extra plugins
This is useful when developing plugins outside Tuleap source tree

The term official is to distinguish from current state of Plugin
implementation in Tuleap. There are 2 kinds of plugins: official and
"is_custom". is_custom were meant to be plugins developed by third party,
outside tuleap tree.

As of today (ie. in a packaged world) this difference is little relevant and
there is no good reason to distinghish the 2 worlds. However it's interesting
to provide a mecanism where you can develop plugins, not integrated in tuleap
sources (or not yet) without having to polluate you worktree (ie. not commited
folder in your working copy, etc).

Change-Id: Ida2ec1f3dcef8aa622fda186332335ac20ece56a