request #8305 System_check event creates missing docman folders
This commit allows the system_check events to create missing folders
for the docman plugin.

In some case, the folder creation does not seem to work.
So the system_check will reod this to ensure that the plugin will works well
for the project.

The same thing is done for SVN or CVS for instance.

Change-Id: I78061430a53e7afc11a02555e7b2487d7d7e70b8