request #8313 Remove spaces at the end of an item title
When submitting an item (a folder for instance) in the docman service
with one or more space at the end (i.e "My title "),
this will lead to an error while accessing this item through Webdav.

So, the spaces at the end will be removed while saving the item.
This behaviour is also implemented for the edition.

Finally, some SOAP automated tests are added for docman plugin.

Change-Id: Ia335169c56d5ad5693cbb39b9f3c2657234e3e84