request #8309: Get rid of PrototypeJS (is-at-top)
Rewrite is-at-top to be jQuery compatible.

is-at-top is used to style the navbar depending on the scrolling
position of the whole page. The classname "is-at-top" should be added to
the body element when the user has not scrolled the page.

Furthermore it fix cross-browser compatibility[1] by using pageYOffset
instead of scrollY property.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Window/scrollY

Change-Id: I7f248fb6dbf532a9dda36d94cf24c32009dd07ce