request #5948 Hiding a value in a select box also hides it in the filter area
Currently if you hide a value, you can no longer filter on that value.

The use case would be for example in a Bug tracker you have versions,
you hide the older versions because those versions are considered end of life but there are bugs that were created on those version.

You want to prevent new artifacts being submitted against those values but should be able to filter.

Change-Id: I9691f2d0881f66c57368fba57f9876dfd5207352