request #8359: tv3->tv5 migration failed when tv3 text field had been switched to numeric ones
In a tracker V3 (legacy), when I transform a text field into a numeric one (int or float)
with histoty activated for ths field, the tv3 to tv5 export process does not work
because numeric fields have some text value (not validated by rnc).

For now, we convert these text values in history to 0 (default values).

Change-Id: Ie85699a401ac98389a6ac7e5102ddf81bfd80a63