request #8360 REST call slowness
Switch Restler to cache routes by default.

This means that, unless you are in DEBUG_MODE, routes will be generated only
once in (var/tmp/tuleap_cache) and if you happen to install new plugins, develop
new routes, etc you will have to manually flush the cache.

Hence, introduce a tool that aimes to ease administrative tasks on Tuleap server

$> tuleap

and, for the time being it's most (and unique) value:

$> tuleap --clear-caches

For production servers it will be automatically deployed in /usr/bin, for devs
you will have to do it manually.

Change-Id: I222885846391da1010c6afb791819bc05ed9f92e