Remove specific management of test_executions
Until now test executions and test definitions where merged because we don't
really advertise test_exec to end user.

However it introduce an unnecessary complexity in both REST routes and angular
frontal to manage this edge case.

We might find a better way to present/hide this complexity later on.

Change-Id: I660602006e27c22999febe64f2c295530ea587b3