request #8375 Give the possibility to complete reference sidebar for an artifact
As of today, we have only references in the sidebar of an artifact.
We would like to have a complement information.
So this changes allow to each plugins can catch the event "TRACKER_EVENT_COMPLEMENT_REFERENCE_INFORMATION" to add informations in the array 'reference_information' structured as :
reference_information = array(
	'title' => (STRING),
	'links' => array(
		'icon' => (STRING),
		'link' => (STRING),
		'label' => (STRING)
	)
)

Change-Id: I91873746f88c3c683f5996f72c4e3ff3ebe1b856