story #8337: List only services activated
The truncated emails' list display only the activated services
in the project that can send truncated emails.
This also fiw the fatal error raised when SVN or FRS is not enable.

In addition, when FRS and SVN are not activated, the checkbox cannot be changed
and a message is displayed to say that no impacted services are available.

Change-Id: I2d34c081972524999889d01b1cfb6e495db906ff