Automatically use sys_default_domain if no adress specified for node server
This commit covers the case where both Tuleap and trafficlights are
behind the same reverse proxy, which is our development setup. Instead
of having to specify the node server adress, we assume that if it
missing it's because we are behind the same reverse proxy, thus using
the sys_default_domain configuration variable.

Change-Id: Ib45e5a8143a76d8adefc11419d080f8dc313dc72