request #8397 SystemEvent_SERVICE_USAGE_SWITCH always in RUNNING status
When a project activates CVS and the package tuleap-core-cvs is not installed,
the system event SystemEvent_SERVICE_USAGE_SWITCH is always in RUNNING state
and blocks other events to be processed.

This case appears because the first event begins to create some folders
and then throws an error (CVS command is missing). However, as this beginning of the
tree is here, the new events will assume that CVS for this project is OK, which is wrong.

This commit adds new log to understand what's going on, and catch a case that
previously leaves the event to RUNNING status

Change-Id: I723c9ceb4331e7ba6b67d5fe8a230069b5567f04